[T] blog.views: delete Methodview replace Blueprint

+++ blog/views.py
@@ -1,12 +1,14 @@
 # -*- coding: utf-8 -*-
 
-from flask import request, jsonify
+from flask import request, Blueprint
 from flask.views import MethodView
 from .models import Post
-from sasukekun_flask.app import app
+from sasukekun_flask import utils
 
-@app.route('/webAPI/test1', methods=['GET', 'POST'])
-def postList():
+blog = Blueprint('blog', __name__)
+
+@blog.route(utils.base_url('/posts/'), methods=['GET', 'POST'])
+def postListAndCreateBlog():
     if request.method == 'GET':
         posts = Post.objects.all()
         category = request.args.get('category')
@@ -19,12 +21,8 @@ def postList():
             posts = posts.filter(tag=tag)
 
         data = [post.to_dict() for post in posts]
-        post_dict = {
-            "code": 0,
-            "data": data,
-            "info": 'ok'
-        }
-        return jsonify(post_dict)
+        return utils.format_response(data=data)
+
     elif request.method == 'POST':
         '''
         Send a json data as follow will create a new blog instance
@@ -39,114 +37,63 @@ def postList():
             "tags": ["tag1", "tag2"]
         }
         '''
-
         data = request.get_json()
 
-        article = Post()
-        article.title = data.get('title')
-        article.slug = data.get('slug')
-        article.abstract = data.get('abstract')
-        article.raw = data.get('raw')
-        article.author = data.get('author')
-        article.category = data.get('category')
-        article.tags = data.get('tags')
+        try:
+            post = Post.objects.get(slug=data.get('slug'))
+            return utils.format_response(code=409, info='This blog exists.')
+        except Post.DoesNotExist:
+            article = Post()
+            article.title = data.get('title')
+            article.slug = data.get('slug')
+            article.abstract = data.get('abstract')
+            article.raw = data.get('raw')
+            article.author = data.get('author')
+            article.category = data.get('category')
+            article.tags = data.get('tags')
 
-        article.save()
+            article.save()
 
-        post_dict = {
-            "code": 0,
-            "data": article.to_dict(),
-            "info": 'ok'
-        }
-
-        return jsonify(post_dict);
-
-class PostListCreateView(MethodView):
-    def get(self):
-        posts = Post.objects.all()
-        category = request.args.get('category')
-        tag = request.args.get('tag')
-
-        if category:
-            posts = posts.filter(category=category)
-
-        if tag:
-            posts = posts.filter(tag=tag)
-
-        data = [post.to_dict() for post in posts]
-
-        return jsonify(posts=data)
-
-    def post(self):
-        '''
-        Send a json data as follow will create a new blog instance
-
-        {
-            "title": "Title1",
-            "slug": "title-1",
-            "abstract": "Abstract for this article",
-            "raw": "The article content",
-            "author": "geezer.",
-            "category": "default",
-            "tags": ["tag1", "tag2"]
-        }
-        '''
-
-        data = request.get_json()
-
-        article = Post()
-        article.title = data.get('title')
-        article.slug = data.get('slug')
-        article.abstract = data.get('abstract')
-        article.raw = data.get('raw')
-        article.author = data.get('author')
-        article.category = data.get('category')
-        article.tags = data.get('tags')
-
-        article.save()
-
-        return jsonify(article.to_dict())
+            return utils.format_response(data=article.to_dict())
 
 
-
-class PostDetailGetUpdateDeleteView(MethodView):
-    def get(self, slug):
+@blog.route(utils.base_url('/posts/<slug>/'), methods=['GET', 'PUT', 'PATCH', 'DELETE'])
+def postDetail(slug):
+    if request.method == 'GET':
         try:
             post = Post.objects.get(slug=slug)
         except Post.DoesNotExist:
-            return jsonify({'error': 'post does not exist'}), 404
+            return utils.format_response(code=404, info='post does not exist')
+        return utils.format_response(data=post.to_dict())
 
-        return jsonify(post=post.to_dict())
-
-    def put(self, slug):
+    elif request.method == 'PUT':
         try:
             post = Post.objects.get(slug=slug)
         except Post.DoesNotExist:
-            return jsonify({'error': 'post does not exist'}), 404
+            return utils.format_response(code=404, info='post does not exist')
 
         data = request.get_json()
 
         if not data.get('title'):
-            return 'title is needed in request data', 400
+            return utils.format_response(code=400, info='title is needed in request data')
 
         if not data.get('slug'):
-            return 'slug is needed in request data', 400
+            return utils.format_response(code=400, info='slug is needed in request data')
 
         if not data.get('abstract'):
-            return 'abstract is needed in request data', 400
+            return utils.format_response(code=400, info='abstract is needed in request data')
 
         if not data.get('raw'):
-            return 'raw is needed in request data', 400
+            return utils.format_response(code=400, info='raw is needed in request data')
 
         if not data.get('author'):
-            return 'author is needed in request data', 400
+            return utils.format_response(code=400, info='author is needed in request data')
 
         if not data.get('category'):
-            return 'category is needed in request data', 400
+            return utils.format_response(code=400, info='category is needed in request data')
 
         if not data.get('tags'):
-            return 'tags is needed in request data', 400
-
+            return utils.format_response(code=400, info='tags is needed in request data')
 
         post.title = data['title']
         post.slug = data['slug']
@@ -158,13 +105,13 @@ class PostDetailGetUpdateDeleteView(MethodView):
 
         post.save()
 
-        return jsonify(post=post.to_dict())
+        return utils.format_response(data=post.to_dict())
 
-    def patch(self, slug):
+    elif request.method == 'PATCH':
         try:
             post = Post.objects.get(slug=slug)
         except Post.DoesNotExist:
-            return jsonify({'error': 'post does not exist'}), 404
+            return utils.format_response(code=404, info='post does not exist')
 
         data = request.get_json()
 
@@ -176,14 +123,13 @@ class PostDetailGetUpdateDeleteView(MethodView):
         post.category = data.get('category') or post.category
         post.tags = data.get('tags') or post.tags
 
-        return jsonify(post=post.to_dict())
+        return utils.format_response(data=post.to_dict())
 
-    def delete(self, slug):
+    elif request.method == 'DELETE':
         try:
             post = Post.objects.get(slug=slug)
         except Post.DoesNotExist:
-            return jsonify({'error': 'post does not exist'}), 404
-
+            return utils.format_response(code=404, info='post does not exist')
         post.delete()
 
-        return 'Succeed to delete post', 204
+        return utils.format_response(data='Succeed to delete post')
